pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "backend"
version = "0.1.0"
description = "Telegram Language bot"
authors = ["Ivan Loboda <loboda.ivan.y@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.32.3"
psycopg2-binary = "^2.9.9"
redis = {extras = ["hiredis"], version = "^5.0.8"}
python-telegram-bot = "^21.6"
langcodes = "3.4.0"
nltk = "3.8.1"
gpt4all = "2.6.0"
pydantic = "^2.9.2"
